+++ src/utils.ts
@@ -2,11 +2,11 @@ interface NotAnArrayType<_T> {
   readonly _: unique symbol
 }
 
-export type ElemType<A> = IsOptional<A> extends true
-  ? NotAnArrayType<A>
-  : A extends (infer Item)[]
-  ? Item
-  : NotAnArrayType<A>
+export type ElemType<A> = IfElse<
+  IsOptional<A>,
+  NotAnArrayType<A>,
+  A extends (infer Item)[] ? Item : NotAnArrayType<A>
+>
 
 // Conduct the check through return value types to work around the
 // Distributive Conditional Types feature:
@@ -28,3 +28,19 @@ type ExtendsNull<A> = Eq<A | null, A>
 type Or<A extends true | false, B extends true | false> = A extends true
   ? true
   : B
+
+export type IfElse<
+  Condition extends true | false,
+  Then,
+  Else
+> = Condition extends true ? Then : Else
+
+export type RequireString<A, B> = IfElse<
+  Eq<A, string>,
+  B,
+  ExpectedStringButGot<A>
+>
+
+interface ExpectedStringButGot<_T> {
+  readonly _: unique symbol
+}